refactor: move infrastructure helper test to right place
